requirements.txt
absl-py==1.0.0
aioconsole==0.3.2
aiohttp==3.7.3
altair==4.2.0
altgraph==0.17
apipkg==1.5
appdirs==1.4.4
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
args==0.1.0
arrow==1.1.0
asgiref==3.3.4
asn1crypto==1.4.0
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
asyncio==3.4.3
asyncio-dgram==2.1.0
atomicwrites==1.4.0
attrs==20.3.0
Automat==20.2.0
autopep8==1.6.0
backcall==0.2.0
base45==0.4.3
beautifulsoup4==4.9.3
black==21.12b0
bleach==3.2.2
blinker==1.4
bs4==0.0.1
build==0.7.0
bump2version==1.0.1
bumpversion==0.6.0
cachetools==4.2.2
cbor2==5.4.2
certifi==2021.5.30
certvalidator==0.11.1
cffi==1.14.6
chardet==4.0.0
clarifai==2.6.2
click==7.1.2
clint==0.5.1
colorama==0.4.4
colorthief==0.2.1
comtypes==1.1.9
configparser==3.8.1
constantly==15.1.0
contextlib2==0.6.0.post1
cose==0.9.dev8
crayons==0.4.0
create-package==1.0.2
cryptography==36.0.1
cssselect==1.1.0
cx-Freeze==6.10
cx-Logging==3.0
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
discord==1.0.1
discord-py-interactions==3.0.2
discord-py-slash-command==3.0.1
discord-ui==5.1.6
discord-webhook==0.14.0
discord.py==1.7.3
discord.py-self==1.9.1
Django==3.2.3
dnspython==2.0.0
docutils==0.16
EasyProcess==1.1
ecdsa==0.17.0
electron-inject==0.7
emoji==1.6.3
engineering-notation==0.6.0
entrypoints==0.3
et-xmlfile==1.1.0
exchange==0.3
execnet==1.8.0
executing==0.8.2
fake-useragent==0.1.11
fastjsonschema==2.15.3
ffmpeg-python==0.2.0
ffpyplayer==4.3.2
Flask==1.1.2
flatbuffers==2.0
fortnite==0.0.8
future==0.18.2
gast==0.4.0
geocoder==1.38.1
giphypop==0.3
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
google-spreadsheet==0.0.6
googleapis-common-protos==1.55.0
googlefinance==0.7
greenlet==1.1.2
grpcio==1.43.0
gTTS==2.2.2
h11==0.12.0
h2==3.2.0
h5py==3.6.0
hpack==3.0.0
httplib2==0.20.4
huepy==1.2.1
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
imageio==2.16.1
imageio-ffmpeg==0.4.5
importlib-metadata==4.11.4
importlib-resources==5.7.1
incremental==21.3.0
iniconfig==1.1.1
instabot==0.117.0
instapy==0.1.1
instapy-chromedriver==2.44
ipwhois==1.2.0
ipykernel==6.6.1
ipython==8.0.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itemadapter==0.3.0
itemloaders==1.0.4
itsdangerous==1.1.0
jaraco.context==4.0.0
jedi==0.18.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.5.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
keyring==22.0.0
lazyme==0.0.23
libclang==12.0.0
lief==0.12.1
lxml==4.6.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mcstatus==6.5.0
MeaningCloud-python==2.0.0
mistune==0.8.4
mock==4.0.3
more-itertools==8.7.0
MouseInfo==0.1.3
moviepy==1.0.3
mplayer.py==0.7.2.post1
mss==6.1.0
multidict==5.1.0
mypy-extensions==0.4.3
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.4
nltk==3.6.5
notebook==6.4.11
numpy==1.20.0
oauth2client==4.1.3
oauthlib==3.1.1
opencv-python==4.5.1.48
openpyxl==3.0.10
opt-einsum==3.3.0
oscrypto==1.2.1
outcome==1.1.0
packaging==20.8
pandas==1.2.3
pandocfilters==1.5.0
parse==1.19.0
parsel==1.6.0
parso==0.8.3
path==15.1.2
path.py==12.5.0
pathspec==0.9.0
pefile==2019.4.18
pep517==0.11.0
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.7.0
platformdirs==2.4.1
playsound==1.2.2
playwright==1.16.0
pluggy==0.13.1
plyer==2.0.0
priority==1.3.0
proglog==0.1.9
prometheus-client==0.14.1
prompt-toolkit==3.0.24
Protego==0.1.16
protobuf==3.19.1
psutil==5.8.0
pure-eval==0.2.1
py==1.10.0
pyarmor==7.3.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/simoe/Downloads/PyAudio-0.2.11-cp38-cp38-win_amd64.whl
PyAutoGUI==0.9.52
pycodestyle==2.8.0
pycparser==2.20
pycryptodome==3.11.0
pydeck==0.7.1
PyDispatcher==2.0.5
pydub==0.25.1
pyee==8.2.2
PyGetWindow==0.0.9
pyglet==1.5.21
Pygments==2.7.4
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
Pympler==1.0.1
PyMsgBox==1.0.9
PyNaCl==1.4.0
pynput==1.7.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
pyppeteer==0.2.6
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyquery==1.4.3
PyRect==0.1.4
pyrsistent==0.18.1
PyScreeze==0.1.26
PySimpleGUI==4.39.1
PySocks==1.7.1
pytest==6.2.3
pytest-shutil==1.7.0
python-dateutil==2.8.1
python-dotenv==0.19.0
python-imagesearch==1.2.0
python-telegram-bot==13.7
python-time==0.3.0
python-vlc==3.0.12118
python3-xlib==0.15
pyttsx3==2.90
PyTweening==1.0.3
pytz==2021.1
PyVirtualDisplay==3.0
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==2.0.5
PyYAML==6.0
pyzbar==0.1.9
pyzmq==22.3.0
qrcode==7.3.1
queuelib==1.6.2
ratelim==0.1.6
readme-renderer==28.0
regex==2021.10.23
requests==2.25.1
requests-html==0.10.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
responses==0.18.0
rfc3986==1.4.0
riotwatcher==3.2.0
rsa==4.8
schedule==1.1.0
scikit-learn==1.0.2
scipy==1.7.3
scrapingant-client==0.3.7
Scrapy==2.5.0
selenium==4.0.0b4
semver==2.13.0
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2
spark-parser==1.8.9
SpeechRecognition==3.8.1
sqlparse==0.4.1
stack-data==0.1.3
stdiomask==0.0.6
streamlit==1.9.0
sty==1.0.0rc2
telegram==0.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.0.0
tinycss2==1.1.1
tk-tools==0.15.0
to-ascii==3.5.8
toml==0.10.2
tomli==1.2.1
toolz==0.11.2
topggpy==1.3.0
tornado==6.1
tqdm==4.56.0
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
twine==3.3.0
Twisted==21.7.0
twisted-iocpsupport==1.0.1
typing_extensions==4.0.1
tzlocal==2.1
uncompyle6==3.8.0
undetected-chromedriver==3.0.3
uritemplate==4.1.1
urllib3==1.26.6
useless-math @ file:///C:/Users/simoe/OneDrive/Documenti/12112
validators==0.19.0
video-to-ascii==1.2.9
w3lib==1.22.0
watchdog==2.1.8
wcwidth==0.2.5
webcolors==1.11.1
webdriver-manager==3.5.1
webencodings==0.5.1
websocket-client==1.2.3
websockets==9.1
Werkzeug==1.0.1
widgetsnbextension==3.6.0
wikipedia==1.4.0
windows-curses==2.2.0
wolframalpha==5.0.0
wrapt==1.13.3
wsproto==1.0.0
xdis==6.0.3
xmltodict==0.12.0
xtermcolor==1.3
xvfbwrapper==0.2.9
yarl==1.6.3
zipp==3.6.0
zope.interface==5.4.0
